Modify modifyWinSize fn to use previous winSize start/end values if possible
If you have winSize set to 'fit' and your List gets hidden, the winSize
is calculated to be 0 while hidden.  Going from 0 back to your previous
size will sometimes modify the start and end bounds of the window which
might appear funny if toggling sizes.  This commit keeps the bounds the
same IF POSSIBLE on resizes from 0 to greater than 0.